backend/requirements.txt
python-dotenv>=1.0.0
langgraph>=0.2.0
langchain-core>=0.3.0
langchain-community>=0.3.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.9.0
pydantic-settings>=2.6.0
httpx>=0.27.0
pytest>=8.3.0
pytest-asyncio>=0.24.0
fastmcp
httpx
primer3-py